Fix DB restore error
restore.sh used DB_USER to create an empty DB if needed.  This failed
but AWS RDS ended up doing the right thing anyway.  The script now uses
DB_ROOTUSER and the default postgres DB when creating the target DB.
The application was upgraded to use Alpine 3.15 and the PostgreSQL 12
client to be compatible with previous backup files.